api_gateway/pyproject.toml
[project]
name = "lifesync-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "vladyslav",email = "llpvo2602ll@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "pydantic (>=2.13.4,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "httpx (>=0.27.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
